wip: add search results screen
